feat: Change amount format from float (currency units) to int64 (cents)
- Change all amount fields from *float64 to *int64 (in cents)
- Update Amount.UnmarshalJSON to handle both number and string types
- Support API returning amounts as numbers or strings
- Update test cases to use cents (e.g., 222.00 USD = 22200 cents)
- Update README with amount format documentation and examples
- All tests passing

+++ model/common/auth_amount.go
@@ -4,8 +4,8 @@ package common
 // Supports: orderAmount, pricingCurrency only
 // Used for: Auth, ForcedAuth, IncrementalAuth
 type AuthAmount struct {
-	// OrderAmount is the order amount (required)
-	OrderAmount *float64 `json:"orderAmount"`
+	// OrderAmount is the order amount in cents (required)
+	OrderAmount *int64 `json:"orderAmount"`
 
 	// PricingCurrency is the pricing currency (ISO 4217, required)
 	PricingCurrency string `json:"pricingCurrency"`
