+++ model/common/batch_total_amount.go
@@ -5,7 +5,7 @@ type BatchTotalAmount struct {
 	// PriceCurrency is the price currency (ISO 4217)
 	PriceCurrency string `json:"priceCurrency"`
 
-	// Amount is the total amount
-	Amount *float64 `json:"amount"`
+	// Amount is the total amount in cents
+	Amount *int64 `json:"amount"`
 }
