+++ model/request/tip_adjust.go
@@ -17,8 +17,8 @@ type TipAdjustRequest struct {
 	// OriginalTransactionRequestID is the original transaction request ID to adjust tip. Either originalTransactionId or originalTransactionRequestId is required. If both are provided, originalTransactionId takes priority
 	OriginalTransactionRequestID string `json:"originalTransactionRequestId,omitempty"`
 
-	// TipAmount is the new tip amount after adjustment, in basic currency unit
-	TipAmount *float64 `json:"tipAmount"`
+	// TipAmount is the new tip amount after adjustment, in cents
+	TipAmount *int64 `json:"tipAmount"`
 
 	// Attach is additional data, returned as-is, recommended to use JSON format
 	Attach string `json:"attach,omitempty"`
